fix: Allow property names with spaces
BREAKING CHANGE: Although this should not break anything, I should like to mention it. Earlier, a space in a property name failed the parsing. Now the parsing will continue, accepting the property name. There was no reason for forbidding spaces in property names. This may change the error message if the source is cut before ending the last property declaration. But the parsing did and will fail in this case anyway.