fix: Allow section names with spaces
BREAKING CHANGE: Although this should not break anything, I should like to mention it. Earlier, a space in a section name failed the parsing. Now the parsing will continue, accepting the section name. There was no reason for forbidding spaces in section names. This may change the error message if the source ends in the middle of a section name. But the parsing did and will fail in this case anyway.